+++ app/setup/executor.go
@@ -210,7 +210,7 @@ func (e *Executor) createComponentStepsByString(componentClient componentEcoSyst
 	}
 
 	result = append(result, component.NewInstallComponentStep(componentClient, name, attributes, namespace))
-	result = append(result, component.NewWaitForComponentStep(componentClient, createComponentLabelSelector(name), namespace, component.DefaultComponentWaitTimeOut5Minutes))
+	result = append(result, component.NewWaitForComponentStep(componentClient, createComponentLabelSelector(name), namespace))
 
 	return result, nil
 }
@@ -250,7 +250,7 @@ func (e *Executor) createLonghornSteps(componentsClient componentEcoSystem.Compo
 	if containsLonghorn {
 		installStep := component.NewInstallComponentStep(componentsClient, longhornComponentName, longhornComponentAttributes, namespace)
 		selector := createComponentLabelSelector(longhornComponentName)
-		waitStep := component.NewWaitForComponentStep(componentsClient, selector, namespace, component.DefaultComponentWaitTimeOut5Minutes)
+		waitStep := component.NewWaitForComponentStep(componentsClient, selector, namespace)
 		result = append(result, installStep)
 		result = append(result, waitStep)
 		delete(components, longhornComponentName)
@@ -270,7 +270,7 @@ func (e *Executor) createComponentSteps(componentsClient componentEcoSystem.Comp
 
 	for componentName, componentAttributes := range e.SetupContext.AppConfig.Components {
 		componentSteps = append(componentSteps, component.NewInstallComponentStep(componentsClient, componentName, componentAttributes, namespace))
-		waitSteps = append(waitSteps, component.NewWaitForComponentStep(componentsClient, createComponentLabelSelector(componentName), namespace, component.DefaultComponentWaitTimeOut5Minutes))
+		waitSteps = append(waitSteps, component.NewWaitForComponentStep(componentsClient, createComponentLabelSelector(componentName), namespace))
 	}
 
 	return componentSteps, waitSteps